pyproject.toml
[tool.poetry]
name = "library-test-task"
version = "0.1.0"
description = ""
authors = ["Aleksandr Rol <rolaleksandr75@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-bootstrap5 = "^24.2"
drf-yasg = "^1.21.7"
django-simple-history = "^3.7.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
